Modify style utils #10
Modify style utils
